+++ app/routes/components/Input/changelog.codegen.ts
@@ -5,4 +5,5 @@ export const changelog = ''
   + '  `--input-placeholder-color`).\n'
   + '- 2025-09-22: Support the `size` props (`sm`, `md`, `lg`) for Input component to\n'
   + '  adjust height and font size. Etc.\n'
+  + '- 2025-09-26: Use theme tokens for Input component styles.\n'
   ;