+++ src/components/Input/index.ts
@@ -1,4 +1,4 @@
-// Input from copy-ui @ 2025-09-22
+// Input from copy-ui @ 2025-09-26
 
 export { Input } from './Input';
 export type { InputProps } from './Input';